apenas falta implementar pydantic e dar update da documentation
