+++ generators/_common.py
@@ -6,6 +6,7 @@ Aqui ficam todas as funções utilitárias que mais do que um gerador usa:
 - linha horizontal
 - carregamento do mapping de produtos (products_mapping.xlsx)
 - extração da quantidade por palete consoante o tipo de cliente (AM/FR/OUTRO)
+- helpers de construção de documentos Word (parágrafos, tabela de datas, cabeçalho)
 """
 
 from pathlib import Path
@@ -14,6 +15,12 @@ from docx.shared import Pt, RGBColor
 from docx.enum.text import WD_ALIGN_PARAGRAPH
 from docx.oxml.ns import qn
 from docx.oxml import OxmlElement
+
+
+# --- Constantes Word partilhadas ---
+
+_W = "http://schemas.openxmlformats.org/wordprocessingml/2006/main"
+_INDENT = -993  # twips — expansão para a margem esquerda (igual em todos os geradores)
 
 
 # --- Formatação de texto ---
@@ -135,6 +142,93 @@ def carregar_mapping_acabamento():
             }
     return mapping
 
+
+# --- Helpers de construção de documentos Word ---
+
+def _set_left_indent_twips(p, twips: int):
+    """Define w:ind/@w:left directamente em twips."""
+    pPr = p._p.get_or_add_pPr()
+    for old in pPr.findall(f"{{{_W}}}ind"):
+        pPr.remove(old)
+    ind = OxmlElement("w:ind")
+    ind.set(qn("w:left"), str(twips))
+    pPr.append(ind)
+
+
+def _body_para(doc, text="", bold=False, underline=False, size_pt=26,
+               align=None, space_after_pt=0, space_before_pt=0):
+    """Adiciona um parágrafo ao body com indentação e fonte padrão dos geradores."""
+    p = doc.add_paragraph()
+    if align is not None:
+        p.alignment = align
+    p.paragraph_format.space_before = Pt(space_before_pt)
+    p.paragraph_format.space_after  = Pt(space_after_pt)
+    _set_left_indent_twips(p, _INDENT)
+    if text:
+        run = p.add_run(text)
+        if underline:
+            run.underline = True
+        set_font(run, font_name="Arial Black", size_pt=size_pt, bold=bold)
+    return p
+
+
+def _update_header_order_num(doc, encomenda):
+    """Substitui o número de ordem na célula (0,2) do header."""
+    _day, _month, year = encomenda["data"].split("/")
+    order_num = f" {encomenda['id']:03d}/{year[2:]} {encomenda['cliente_tipo']}"
+
+    header = doc.sections[0].header
+    cell   = header.tables[0].cell(0, 2)
+    para   = cell.paragraphs[0]
+
+    for run in list(para.runs):
+        run._r.getparent().remove(run._r)
+
+    run = para.add_run(order_num)
+    run.bold = True
+    set_font(run, font_name="Arial", size_pt=20, bold=True)
+
+
+def _make_date_table_inline(doc):
+    """Converte a tabela flutuante de datas em inline, alinhando a borda direita
+    com a borda direita do cabeçalho (tblInd=4813 para tblW=4819, header=10632/-1000).
+    """
+    body = doc.element.body
+    for tbl in body.findall(f"{{{_W}}}tbl"):
+        tblPr = tbl.find(f"{{{_W}}}tblPr")
+        if tblPr is None:
+            continue
+        for tag in (f"{{{_W}}}tblpPr", f"{{{_W}}}tblOverlap"):
+            el = tblPr.find(tag)
+            if el is not None:
+                tblPr.remove(el)
+        tblW = tblPr.find(f"{{{_W}}}tblW")
+        if tblW is None:
+            tblW = OxmlElement("w:tblW")
+            tblPr.insert(0, tblW)
+        tblW.set(qn("w:w"), "4819")
+        tblW.set(qn("w:type"), "dxa")
+        tblInd = tblPr.find(f"{{{_W}}}tblInd")
+        if tblInd is None:
+            tblInd = OxmlElement("w:tblInd")
+            tblPr.append(tblInd)
+        tblInd.set(qn("w:w"), "4813")
+        tblInd.set(qn("w:type"), "dxa")
+        jc = tblPr.find(f"{{{_W}}}jc")
+        if jc is None:
+            jc = OxmlElement("w:jc")
+            tblPr.append(jc)
+        jc.set(qn("w:val"), "left")
+
+
+def _clear_body_paragraphs(doc):
+    """Remove todos os parágrafos do body, preservando tabelas e sectPr."""
+    body = doc.element.body
+    for child in [c for c in body if c.tag == f"{{{_W}}}p"]:
+        body.remove(child)
+
+
+# --- Mapping de produtos ---
 
 def qt_palete(qt_str: str, cliente_tipo: str) -> str:
     """Da string '20/30' (AM/FR), devolve a parte que corresponde ao cliente_tipo.
