+++ main.py
@@ -4,6 +4,8 @@ Este ficheiro é deliberadamente curto: só cria a app, monta os ficheiros
 estáticos e regista as rotas. Toda a lógica vive em `app/`.
 """
 
+import shutil
+from datetime import date
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 
@@ -11,9 +13,20 @@ from app import config
 from app.routes import register_all
 
 
+def _backup_historico():
+    """Faz uma cópia diária do historico.json em data/backups/ no arranque."""
+    backup_dir = config.DATA_DIR / "backups"
+    backup_dir.mkdir(exist_ok=True)
+    if config.DATA_FILE.exists():
+        dest = backup_dir / f"historico_{date.today().isoformat()}.json"
+        if not dest.exists():
+            shutil.copy2(config.DATA_FILE, dest)
+
+
 def criar_app() -> FastAPI:
     """Cria e configura a instância FastAPI."""
     config.garantir_pastas()
+    _backup_historico()
 
     app = FastAPI(title="Gestão de Encomendas - Colmeias")
     app.mount("/static", StaticFiles(directory=str(config.STATIC_DIR)), name="static")
